+++ common/compose/src/jvmTest/kotlin/LoadableComponentTests.kt
@@ -2,11 +2,9 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.test.ExperimentalTestApi
 import androidx.compose.ui.test.runComposeUiTest
 import dev.inmo.micro_utils.common.compose.LoadableComponent
-import dev.inmo.micro_utils.coroutines.SpecialMutableStateFlow
-import kotlinx.coroutines.flow.MutableSharedFlow
+import dev.inmo.micro_utils.coroutines.MutableRedeliverStateFlow
 import kotlinx.coroutines.flow.filter
 import kotlinx.coroutines.flow.first
-import kotlinx.coroutines.flow.firstOrNull
 import org.jetbrains.annotations.TestOnly
 import kotlin.test.Test
 import kotlin.test.assertTrue
@@ -16,8 +14,8 @@ class LoadableComponentTests {
     @Test
     @TestOnly
     fun testSimpleLoad() = runComposeUiTest {
-        val loadingFlow = SpecialMutableStateFlow<Int>(0)
-        val loadedFlow = SpecialMutableStateFlow<Int>(0)
+        val loadingFlow = MutableRedeliverStateFlow<Int>(0)
+        val loadedFlow = MutableRedeliverStateFlow<Int>(0)
         setContent {
             LoadableComponent<Int>({
                 loadingFlow.filter { it == 1 }.first()